Added visual touch
Separated protocols from app.py. Added visual overhaul to the frontend.
